backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-decouple==3.8
psycopg[binary]>=3.1.0
Pillow>=11.0.0
requests==2.31.0
python-dotenv==1.0.1
